:wrench: Move button to its own file
